Fixed `remove` by tweaking `uninsert`
Now, `uninsert t` returns a *quadruple* `(w',a',lb,mt')`:

1. `w'`, the weight of the uninserted thing;
2. `a'`, the uninserted thing itself;
3. `lb`, the lower bound of the bucket that `a'` was in; and
4. `mt'`, the new tree.

It's the lower bound that's new.  We need this for `remove`, which takes
an index `i` into the old tree, *not* the new one.  If `mt'` is
`Nothing`, then we're done – we must have had a one-element tree, so the
index pointed there and we've successfully removed it.  Otherwise, if
`mt'` is `Just t'`, there are three cases:

1. If `i < lb`, then the place `i` points to wasn't affected by the
   uninsert; thus, we can just use it as-is to `replace w' a' t' i`.
2. If `i ∈ [lb,lb+w')`, then the thing we uninserted was what we wanted
   to remove; we can return it.
3. Otherwise, if `i >= lb+w'`, then the place `i` points to is now
   wrong; we subtract off `w'` to find the corresponding index into
   `t'` (and so call `replace w' a' t' (i - w')`).